+++ spec/terraform_plan_spec.rb
@@ -65,6 +65,22 @@ describe TerraformLandscape::TerraformPlan do
       OUT
     end
 
+    context 'when output contains a multiple resources with no attributes' do
+      let(:terraform_output) { normalize_indent(<<-TXT) }
+        + some_resource_type.some_resource_name
+
+        + some_resource_type.some_resource_name
+
+      TXT
+
+      it { should == normalize_indent(<<-OUT) }
+        + some_resource_type.some_resource_name
+
+        + some_resource_type.some_resource_name
+
+      OUT
+    end
+
     context 'when output contains a single resource with multiple attributes' do
       let(:terraform_output) { normalize_indent(<<-TXT) }
         ~ some_resource_type.some_resource_name